Strip only the  'highlight=...' query string
There is actually an API to do exactly what we want to do: delete a
search parameter from the query string. By switching to this API, we can
get precisely the desired behavior.

As a minor improvement, the redirect listener now only triggers on URLs
that have "highlight" after the '?' signifying the query string. This
should significantly cut down on the number of requests going through
the listener.

Note that I was unable to find explicit documentation that the `return;`
(aka `return undefined;`) behavior from the listener behaves as
required. However, one of the examples on the Mozilla documentation for
browser.webRequest.onBeforeRequest also returns undefined, so it appears
that this is the supported behavior. If this turns out to be incorrect,
it may be possible to instead use onBeforeRedirect to load this listener
if it would need to redirect, with the listener always unregistering
itself.